+++ example/tests/fixture_to_include.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+"""
+This file only serves as a test fixture to be included.
+"""
+
+__author__ = 'sobolevn'
+
+FIXTURE_VALUE = __file__