+++ split_settings/__init__.py
@@ -6,6 +6,6 @@ Easily override and modify settings. Use wildcards and optional
 settings files.
 """
 
-__version__ = '0.2.0'
+__version__ = '0.2.1'
 
 __all__ = ['__version__']